+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Adds methods `Entity.ID` and `Entity.Gen` for debugging purposes (#134)
 - Adds `MapX.NewEntityFn`, `MapX.AddFn`, `Map.NewEntityFn` and `Map.AddFn` (#145)
 - Adds `ExchangeX.AddFn` and `ExchangeX.ExchangeFn` (#145)
+- Adds `World.Stats` for extracting world statistics (#147)
 
 ### Documentation